+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `MTLResourceID::from_raw` and `MTLResourceID::as_raw` to allow
   querying the underlying data.
 * Added new framework crates:
+  - `AudioToolbox` / `objc2-audio-toolbox`.
   - `CoreAudioTypes` / `objc2-core-audio-types`.
   - `CoreGraphics` / `objc2-core-graphics`.
   - `CoreFoundation` / `objc2-core-foundation`.